splitting util functions
